+++ lib/data/model/res_model/get_all_cart_res_model/get_all_cart_res_model.dart
@@ -81,7 +81,7 @@ class ProductDetails with _$ProductDetails {
     @JsonKey(name: "mainImage")
     String? mainImage,
     @JsonKey(name: "itemsWeight")
-    int? itemsWeight,
+    double? itemsWeight,
     @JsonKey(name: "images")
     List<Image>? images,
     @JsonKey(name: "scales")
